+++ src/components/ToolTips.tsx
@@ -7,7 +7,7 @@ import * as OF from 'office-ui-fabric-react';
 import { FormattedMessage } from 'react-intl'
 import { FM } from '../react-intl-messages'
 import { MemoryValue } from '@conversationlearner/models'
-import HelpLink from '../components/HelpLink'
+import HelpLink from './HelpLink'
 import './ToolTips.css'
 
 export enum TipType {
