+++ src/components/modals/ActionCreatorEditor.tsx
@@ -12,7 +12,7 @@ import actions from '../../actions'
 import { Modal } from 'office-ui-fabric-react/lib/Modal'
 import { ActionBase, ActionTypes, AppBase, EntityBase, EntityType, RenderedActionArgument, SessionAction, TextAction, ApiAction, CardAction, IActionArgument, Callback, Template, ActionPayload, CardPayload, TextPayload } from '@conversationlearner/models'
 import ConfirmCancelModal from './ConfirmCancelModal'
-import EntityCreatorEditor from './EntityCreatorEditor'
+import EntityCreatorEditor from './EntityCreatorEditorBackup'
 import AdaptiveCardViewer from './AdaptiveCardViewer/AdaptiveCardViewer'
 import * as ActionPayloadEditor from './ActionPayloadEditor'
 import { State } from '../../types'
