+++ src/components/modals/TeachSessionModal.tsx
@@ -7,7 +7,7 @@ import './TeachSessionModal.css';
 import { returntypeof } from 'react-redux-typescript';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { ErrorHandler } from './../../ErrorHandler'
+import { ErrorHandler } from '../../ErrorHandler'
 import { AT } from '../../types/ActionTypes'
 import { Modal } from 'office-ui-fabric-react/lib/Modal';
 import * as OF from 'office-ui-fabric-react';
